src/gui/tcltk/tcl:   Ported 8 csets from /home/rob/tcl-upgrade/tcltk-today-with-8_6_b2
bk: 52fcd888pAURtJUdEJpVTjywlEdWOA

+++ src/gui/tcltk/tcl/unix/tclUnixCompat.c
@@ -5,9 +5,6 @@
  *
  * See the file "license.terms" for information on usage and redistribution of
  * this file, and for a DISCLAIMER OF ALL WARRANTIES.
- *
- * RCS: @(#) $Id$
- *
  */
 
 #include "tclInt.h"
@@ -736,7 +733,7 @@ CopyArray(
     p = buf + len;
 
     for (j = 0; j < i; j++) {
-	int sz = (elsize<0 ? strlen(src[j])+1 : elsize);
+	int sz = (elsize<0 ? (int) strlen(src[j]) + 1 : elsize);
 
 	len += sz;
 	if (len > buflen) {
